move PasswordField to standalone module, loaded lazyly
